Third homework source

+++ app/src/main/java/com/example/kimichael/yamblz_forecast/data/network/forecast/response/Wind.java
@@ -7,23 +7,23 @@ import com.google.gson.annotations.SerializedName;
 public class Wind {
 
     @SerializedName("speed")
-    private Integer speed;
+    private Double speed;
     @SerializedName("deg")
-    private Integer deg;
+    private Double deg;
 
-    public Integer getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(Integer speed) {
+    public void setSpeed(Double speed) {
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
